update and organize components, fix build and linting errors
