+++ app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
 // library imports
-import { handlers } from "@/auth";
+import { handlers } from "@/utils/auth";
 
 export const { GET, POST } = handlers;
