+++ components/Assistant/strategies-sidebar.tsx
